Move command-options file to lib
