+++ packages/core/lib/command-options.js
